+++ Packages/com.vareiko.foundation/Runtime/App/IApplicationLifecycleService.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace Vareiko.Foundation.App
+{
+    public interface IApplicationLifecycleService
+    {
+        bool IsPaused { get; }
+        bool HasFocus { get; }
+        bool IsQuitting { get; }
+
+        event Action<bool> PauseChanged;
+        event Action<bool> FocusChanged;
+        event Action QuitRequested;
+    }
+}